MSVC: fix ambiguous call to pow(). There are "float pow( float, int )" and "double pow( double, int )".
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@45782 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ examples/Plotter/BigDataset/MainWidget.cpp
@@ -76,5 +76,5 @@ void MainWidget::addPointsButtonClicked()
 {
     int idx = m_addPointsButtons.indexOf( qobject_cast< QPushButton* >( sender() ) );
     Q_ASSERT( idx >= 0 );
-    m_model.appendPoints( pow( 10, idx + 3 ) );
+    m_model.appendPoints( pow( qreal( 10 ), qreal( idx + 3 ) ) );
 }
